workshop/pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "Valkey Caching Workshop with Flughafen DB SQLModel classes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlmodel>=0.0.14",
    "python-dotenv>=1.0.0",
    "pymysql>=1.1.0",
    "psycopg2-binary>=2.9.0",
]
